tasks up to 4.22.

+++ controllers/blogs.js
@@ -1,8 +1,11 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1, id: 1 })
+
   response.json(blogs)
 })
 
@@ -11,17 +14,23 @@ blogsRouter.get('/:id', async (request, response) => {
   response.json(blogToShow)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
   const body = request.body
+
+  const user = request.user
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes == undefined ? 0 : body.likes
+    likes: body.likes == undefined ? 0 : body.likes,
+    user: user._id
   })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.json(savedBlog)
 })
 
@@ -40,8 +49,16 @@ blogsRouter.put('/:id', async (request, response) => {
   response.json(updatedNote)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() == user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+  } 
+  else {
+    return response.status(401).json({ error: 'You are not allowed to delete this post!' })
+  }
   response.status(204).end()
 })
